+++ tests/Paladin.Tests/Program.cs
@@ -7,5 +7,6 @@ Console.WriteLine("  " + new string('-', 46));
 ShieldTests.Register();
 SessionTests.Register();
 ReplayAndLaunchTests.Register();
+PathDisplayTests.Register();
 
 return TestHarness.Run();